feat: 업스크롤 로직 분리
